change docker push after build to be optional
Previously there was always a docker push in build.sh.
Pushing will remain as default behavior.

With authenticated registry (Nexus) for CI, we cannot push without
a login secret.
Here, we will disable docker push with CONTRAIL_REGISTRY_PUSH=0.
The secret is only present in post-run stage of CI, where we will
do a separate push logic.

Use variable instead of the existence-of-logfile to gather the status.

Change-Id: I06ae8679d57fe1053e46be2644b77cb38e9d1a9a
Partial-Jira-bug: JD-428